Shiro 权限单元测试提交

+++ src/main/java/com/bc/pmpheep/back/dao/WriterPermissionDao.java
@@ -43,9 +43,9 @@ public interface WriterPermissionDao {
      */
     WriterPermission getWriterPermissionByPermissionName(String permissionName);
 
-    Integer delete(int id);
+    Integer delete(Long id);
 
-    WriterPermission get(int id);
+    WriterPermission get(Long id);
 
     List<WriterPermission> getListResource();
 }
