+++ src/main/java/com/bc/pmpheep/back/service/PmphRoleService.java
@@ -27,14 +27,15 @@ public interface PmphRoleService {
      * 
      * @param id
      */
-    Integer delete(int id) throws CheckedServiceException;
+    Integer delete(Long id) throws CheckedServiceException;
 
     /**
+     * 根据角色 id 删除对应资源
      * 
      * @param ids
      * @return
      */
-    void deleteRoleAndResource(List<Integer> ids) throws CheckedServiceException;
+    void deleteRoleAndResource(List<Long> ids) throws CheckedServiceException;
 
     /**
      * 根据 id 加载角色对象
@@ -42,7 +43,7 @@ public interface PmphRoleService {
      * @param id
      * @return
      */
-    PmphRole get(int id) throws CheckedServiceException;
+    PmphRole get(Long id) throws CheckedServiceException;
 
     /**
      * 查询所有角色对象的列表
@@ -72,7 +73,7 @@ public interface PmphRoleService {
      * @param roleId
      * @return
      */
-    PmphUserRole getUserRole(int uid, int roleId) throws CheckedServiceException;
+    PmphUserRole getUserRole(Long uid, Long roleId) throws CheckedServiceException;
 
     /**
      * 根据用户 id 和角色 id 插入一条用户角色关系数据
@@ -80,7 +81,7 @@ public interface PmphRoleService {
      * @param uid
      * @param roleId
      */
-    void addUserRole(int uid, int roleId) throws CheckedServiceException;
+    void addUserRole(Long uid, Long roleId) throws CheckedServiceException;
 
     /**
      * 根据用户 id 和角色 id 删除一条用户角色关系数据
@@ -88,14 +89,14 @@ public interface PmphRoleService {
      * @param uid
      * @param roleId
      */
-    void deleteUserRole(int uid, int roleId) throws CheckedServiceException;
+    void deleteUserRole(Long uid, Long roleId) throws CheckedServiceException;
 
     /**
      * 删除某个用户的所有角色
      * 
      * @param uid
      */
-    void deleteUserRoles(int uid) throws CheckedServiceException;
+    void deleteUserRoles(Long uid) throws CheckedServiceException;
 
     /**
      * 根据角色id获取可以访问的所有资源
@@ -103,7 +104,7 @@ public interface PmphRoleService {
      * @param roleId
      * @return
      */
-    List<PmphPermission> getListRoleResource(int roleId) throws CheckedServiceException;
+    List<PmphPermission> getListRoleResource(Long roleId) throws CheckedServiceException;
 
     /**
      * 根据角色 id 和权限 id 增加一条用户权限关联数据
@@ -111,7 +112,7 @@ public interface PmphRoleService {
      * @param roleId
      * @param resId
      */
-    void addRoleResource(int roleId, int resId) throws CheckedServiceException;
+    void addRoleResource(Long roleId, Long resId) throws CheckedServiceException;
 
     /**
      * 根据角色 id 和权限 id 删除一条用户权限关联数据
@@ -119,7 +120,7 @@ public interface PmphRoleService {
      * @param roleId
      * @param resId
      */
-    void deleteRoleResource(int roleId, int resId) throws CheckedServiceException;
+    void deleteRoleResource(Long roleId, Long resId) throws CheckedServiceException;
 
     /**
      * 根据角色 id 和权限 id 查询一条用户权限关联数据
@@ -128,7 +129,7 @@ public interface PmphRoleService {
      * @param resId
      * @return
      */
-    PmphRolePermission getResourceRole(int roleId, int resId) throws CheckedServiceException;
+    PmphRolePermission getResourceRole(Long roleId, Long resId) throws CheckedServiceException;
 
-    Integer deleteRoleAndUser(List<Integer> ids) throws CheckedServiceException;
+    Integer deleteRoleAndUser(List<Long> ids) throws CheckedServiceException;
 }
