+++ src/main/java/com/bc/pmpheep/back/service/WriterRoleService.java
@@ -27,14 +27,14 @@ public interface WriterRoleService {
      * 
      * @param id
      */
-    Integer delete(int id) throws CheckedServiceException;
+    Integer delete(Long id) throws CheckedServiceException;
 
     /**
      * 
      * @param ids
      * @return
      */
-    void deleteRoleAndResource(List<Integer> ids) throws CheckedServiceException;
+    void deleteRoleAndResource(List<Long> ids) throws CheckedServiceException;
 
     /**
      * 根据 id 加载角色对象
@@ -42,7 +42,7 @@ public interface WriterRoleService {
      * @param id
      * @return
      */
-    WriterRole get(int id) throws CheckedServiceException;
+    WriterRole get(Long id) throws CheckedServiceException;
 
     /**
      * 查询所有角色对象的列表
@@ -72,7 +72,7 @@ public interface WriterRoleService {
      * @param roleId
      * @return
      */
-    WriterUserRole getUserRole(int uid, int roleId) throws CheckedServiceException;
+    WriterUserRole getUserRole(Long uid, Long roleId) throws CheckedServiceException;
 
     /**
      * 根据用户 id 和角色 id 插入一条用户角色关系数据
@@ -80,7 +80,7 @@ public interface WriterRoleService {
      * @param uid
      * @param roleId
      */
-    void addUserRole(int uid, int roleId) throws CheckedServiceException;
+    void addUserRole(Long uid, Long roleId) throws CheckedServiceException;
 
     /**
      * 根据用户 id 和角色 id 删除一条用户角色关系数据
@@ -88,14 +88,14 @@ public interface WriterRoleService {
      * @param uid
      * @param roleId
      */
-    void deleteUserRole(int uid, int roleId) throws CheckedServiceException;
+    void deleteUserRole(Long uid, Long roleId) throws CheckedServiceException;
 
     /**
      * 删除某个用户的所有角色
      * 
      * @param uid
      */
-    void deleteUserRoles(int uid) throws CheckedServiceException;
+    void deleteUserRoles(Long uid) throws CheckedServiceException;
 
     /**
      * 根据角色id获取可以访问的所有资源
@@ -103,7 +103,7 @@ public interface WriterRoleService {
      * @param roleId
      * @return
      */
-    List<WriterPermission> getListRoleResource(int roleId) throws CheckedServiceException;
+    List<WriterPermission> getListRoleResource(Long roleId) throws CheckedServiceException;
 
     /**
      * 根据角色 id 和权限 id 增加一条用户权限关联数据
@@ -111,7 +111,7 @@ public interface WriterRoleService {
      * @param roleId
      * @param resId
      */
-    void addRoleResource(int roleId, int resId) throws CheckedServiceException;
+    void addRoleResource(Long roleId, Long resId) throws CheckedServiceException;
 
     /**
      * 根据角色 id 和权限 id 删除一条用户权限关联数据
@@ -119,7 +119,7 @@ public interface WriterRoleService {
      * @param roleId
      * @param resId
      */
-    void deleteRoleResource(int roleId, int resId) throws CheckedServiceException;
+    void deleteRoleResource(Long roleId, Long resId) throws CheckedServiceException;
 
     /**
      * 根据角色 id 和权限 id 查询一条用户权限关联数据
@@ -128,7 +128,7 @@ public interface WriterRoleService {
      * @param resId
      * @return
      */
-    WriterRolePermission getResourceRole(int roleId, int resId) throws CheckedServiceException;
+    WriterRolePermission getResourceRole(Long roleId, Long resId) throws CheckedServiceException;
 
-    Integer deleteRoleAndUser(List<Integer> ids) throws CheckedServiceException;
+    Integer deleteRoleAndUser(List<Long> ids) throws CheckedServiceException;
 }
